% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_smooth_normalise}
\alias{add_smooth_normalise}
\title{Additive Smoothing and Normalisation}
\usage{
add_smooth_normalise(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\description{
Where possible, the \code{{information}} package relies on microservices architecture. This function serves to increment all values in distributions that include at least on zero and would otherwise lead to log(0) problems with a tiny number, then normalise the resulting distribution.
}
