% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Aggregate.R
\name{Aggregate}
\alias{Aggregate}
\title{(helper function )Aggregate data for WOE/NWOE calculations}
\usage{
Aggregate(data, x, y, breaks, trt)
}
\arguments{
\item{data}{input data frame}

\item{x}{variable to be aggregated}

\item{y}{dependent variable}

\item{breaks}{breaks for binning}

\item{trt}{binary treatment variable (for net lift only)}
}
\description{
\code{Aggregate} returns aggregated data for the WOE and NWOE functions
}

