% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CheckInputs.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Check user inputs and convert factors to characters. Provide readable reasons if errors are found.}
\usage{
CheckInputs(train, valid, trt, y)
}
\arguments{
\item{train}{training data.}

\item{valid}{validation dataset (default is NULL)}

\item{trt}{treatment indicator}

\item{y}{dependent variable}
}
\description{
\code{CheckInputs} Checks user inputs and converts factors to characters. Returns the altered datasets as a list.
Provides readable reasons if errors are found.
}

