% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateTables.R
\name{comb}
\alias{comb}
\title{Helper function for foreach when returning lists with multiple objects
\code{comb} Helper function for foreach when returning lists with multiple objects}
\usage{
comb(x, ...)
}
\description{
Helper function for foreach when returning lists with multiple objects
\code{comb} Helper function for foreach when returning lists with multiple objects
}

