% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateInfoTables.R
\name{CreateInfoTables}
\alias{CreateInfoTables}
\title{Create WOE/NWOE tables and rank variables by IV/NIV}
\usage{
CreateInfoTables(data, valid = NULL, y, bins = 10, trt = NULL,
  ncore = NULL)
}
\arguments{
\item{data}{input dataset for analysis (this is typically your training dataset)}

\item{valid}{validation dataset (default is NULL). Must have the same variables as the training dataset}

\item{y}{dependent variable}

\item{bins}{number of bins (default is 10)}

\item{trt}{binary treatment variable (for net lift modeling -- i.e., NWOE and NIV). Default is NULL (for standard WOE and IV)}

\item{ncore}{number of cores used. Default is to use detectcores-1.}
}
\description{
\code{CreateInfoTables} returns WOE or NWOE tables (as data.frames), and a data.frame with IV or NIV values for all
predictive variables.
}
\examples{
##------------------------------------------------------------
## WOE analysis, no validation
##------------------------------------------------------------
library(Information)
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- Information::CreateInfoTables(data=train, y="PURCHASE", ncore=2)
print(head(IV$Summary), row.names=FALSE)
print(IV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)

##------------------------------------------------------------
## NWOE analysis, no validation
##------------------------------------------------------------
data(train, package="Information")
NIV <- Information::CreateInfoTables(data=train, trt="TREATMENT", y="PURCHASE", ncore=2)
print(head(NIV$Summary), row.names=FALSE)
print(NIV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)

##------------------------------------------------------------
## NWOE analysis, validation
##------------------------------------------------------------
data(train, package="Information")
data(valid, package="Information")
NIV <- Information::CreateInfoTables(data=train,
                                     valid=valid,
                                     trt="TREATMENT",
                                     y="PURCHASE", ncore=2)
print(head(NIV$Summary), row.names=FALSE)
print(NIV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)
}

