% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateTables.R
\name{CreateTables}
\alias{CreateTables}
\title{Create WOE/NWOE tables and rank variables by IV/NIV}
\usage{
CreateTables(data, valid = NULL, y, bins = 10, trt = NULL, ncore = NULL)
}
\arguments{
\item{data}{input dataset for analysis (this is typically your training dataset)}

\item{valid}{validation dataset (default is NULL)}

\item{y}{dependent variable}

\item{bins}{number of bins (default is 10)}

\item{trt}{binary treatment variable (for net lift modeling -- i.e., NWOE and NIV). Default is NULL (for standard WOE and IV)}

\item{ncore}{number of cores used. Default is to use detectcores-1.}
}
\description{
\code{CreateTables} returns WOE or NWOE tables (as data.frames), and a data.frame with IV or NIV values for all
predictive variables.
}
\examples{
##------------------------------------------------------------
## WOE analysis, no validation
##------------------------------------------------------------
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- CreateTables(data=train, y="PURCHASE")
print(head(IV$Summary), row.names=FALSE)
print(IV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)

##------------------------------------------------------------
## NWOE analysis, no validation
##------------------------------------------------------------
data(train, package="Information")
NIV <- CreateTables(data=train, trt=TREATMENT, y="PURCHASE")
print(head(NIV$Summary), row.names=FALSE)
print(NIV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)

##------------------------------------------------------------
## NWOE analysis, validation
##------------------------------------------------------------
data(train, package="Information")
data(valid, package="Information")
NIV <- CreateTables(data=train, valid=valid, trt=TREATMENT, y="PURCHASE")
print(head(NIV$Summary), row.names=FALSE)
print(NIV$Tables$N_OPEN_REV_ACTS, row.names=FALSE)
}

