% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epsilon}
\alias{epsilon}
\title{Output a tiny number to avoid log(0)}
\usage{
epsilon()
}
\description{
Where possible, the \code{{information}} package relies on microservices architecture. This function serves only one purpose: to output a tiny value. This is necessary to replace zeroes and avoid log(0), which is undefined, in some cases. By having a single function do this, it is easier to keep things consistent throughout the package.
}
