% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IsBinary.R
\name{is.binary}
\alias{is.binary}
\title{(helper function) Calculate cross validation penalty}
\usage{
is.binary(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
\code{is.binary} returns TRUE if a variable is binary, and FALSE otherwise
}

