% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPlot.R
\name{MultiPlot}
\alias{MultiPlot}
\title{(helper function) Plot mutiple WOE vectors on one page}
\usage{
MultiPlot(information_object, variables, same_scales = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variables}{vector of variables that you want to compare}

\item{same_scales}{set to TRUE if all plots should have the same limits on the y-axes (default is FALSE)}
}
\description{
\code{MultiPlot} creates a multiple WOE or NWOE bar charts on the same page for a specified vector of variables.
}

