% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiWOEPlot.R
\name{MultiPlotWOE}
\alias{MultiPlotWOE}
\title{Plot mutiple WOE vectors on one page}
\usage{
MultiPlotWOE(information_object, variables, same_scale = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variables}{vector of variables that you want to compare}

\item{same_scale}{should all plots have the same limits on the y-axes (default is FALSE)}
}
\description{
\code{MultiPlotWOE} creates a multiple WOE bar charts on the same page for a specified vector of variables.
Multiple pages will be created if needed (9 plots per page).
}
\examples{
##------------------------------------------------------------
## WOE plots
##------------------------------------------------------------
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- CreateTables(data=train, y="PURCHASE")
# Plotting two variables
MultiPlotWOE(IV, c("N_OPEN_REV_ACTS", "AGE"))
# Plotting all variables with IV>0.1
MultiPlotWOE(IV, subset(IV$Summary, IV>0.1)$Variable)
}

