% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NWOE.R
\name{NWOE}
\alias{NWOE}
\title{Create WOE table}
\usage{
NWOE(t, x)
}
\arguments{
\item{t}{table prepared by the Aggregate function}

\item{x}{predictor}
}
\description{
\code{WOE} returns NWOE tables from data frames prepared by Aggregate(). This is only for net lift models.
}

