% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Penalty.R
\name{penalty}
\alias{penalty}
\title{(helper function) Calculate cross validation penalty}
\usage{
penalty(t, v, d_net_lift)
}
\arguments{
\item{t}{training data (data.frame)}

\item{v}{valdation data (data.frame)}

\item{d_net_lift}{is it a net lift model? (1=yes, 0=no)}
}
\description{
\code{penalty} returns the weighted cross validation penalty.
}

