% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_infotables.R
\name{plot_infotables}
\alias{plot_infotables}
\title{Plot WOE or NWOE vectors}
\usage{
plot_infotables(information_object = NULL, variables = NULL,
  same_scales = FALSE, show_values = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variables}{vector of one more variables}

\item{same_scales}{should all plots have the same limits on the y-axes (default is FALSE)}

\item{show_values}{if set to TRUE, values will be displayed on the bar chart (default is FALSE)}
}
\description{
\code{plot_infotable} creates WOE or NWOE bar charts for one or more variables.
For multi-variable plots, bar charts will be displayed in a grid
If more than 9 variables are plotted, plot_infotable will create multiple pages
}
\examples{
##------------------------------------------------------------
## WOE plots
##------------------------------------------------------------
library(Information)
data(train, package="Information")
train <- subset(train, TREATMENT==1)[1:1000,]
IV <- Information::create_infotables(data=train, y="PURCHASE", parallel=FALSE)

# Plotting a single variable
Information::plot_infotables(IV, "N_OPEN_REV_ACTS")

# Plotting multiple variables
Information::plot_infotables(IV, IV$Summary$Variable[1:4], same_scale=TRUE)

closeAllConnections()
}

