% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SinglePlot.R
\name{SinglePlot}
\alias{SinglePlot}
\title{(helper function) Plot a WOE or NWOE vector}
\usage{
SinglePlot(information_object, variable, show_values = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variable}{variable for which we want to see the WOE pattern}

\item{show_values}{if set to TRUE, values will be displayed on the bar chart (default is FALSE)}
}
\description{
\code{SinglePlot} creates a bar chart of the WOE or NWOE pattern for a specified variable
}

