% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train}
\alias{train}
\title{Training dataset}
\format{A data frame with 4972 rows and 70 variables:
\itemize{
  \item TREATMENT: equals 1 if the person received the marketing offer, and 0 if the person was in the control group
  \item PURCHASE: equals 1 if the person accepted the offer, and 0 otherwise
  \item UNIQUE_ID: unique identifier
  \item Predictive variables include demographic, geographic and credit bureau data
}}
\usage{
train
}
\description{
The data is from a historical marketing campaign.
It contains 68 predictive variables and 4,972 records. In addition, it contains a
treatment indicator and a purchase indicator.
}
\keyword{datasets}

