% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateTables.R
\name{CreateTables}
\alias{CreateTables}
\title{Create WOE/NWOE tables and rank variables by IV/NIV}
\usage{
CreateTables(data, valid = NULL, y, bins = 10, trt = NULL, ncore = NULL)
}
\arguments{
\item{data}{input dataset for analysis (this is typically your training dataset)}

\item{valid}{validation dataset (default is NULL)}

\item{y}{dependent variable}

\item{bins}{number of bins (default is 10)}

\item{trt}{binary treatment variable (for net lift modeling). Default is NULL}

\item{ncore}{number of cores used. Default is to use detectcores-1.}
}
\description{
\code{CreateTables} returns WOE or NWOE tables (as data.frames), and a data.frame with IV or NIV values for all
predictive variables.
}
\examples{
##------------------------------------------------------------
## WOE analysis, no validation
##------------------------------------------------------------
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- CreateTables(data=, y=PURCHASE)
View(IV$Summary)
IV$Tables$N_OPEN_REV_ACTS

##------------------------------------------------------------
## NWOE analysis, no validation
##------------------------------------------------------------
data(train, package="Information")
NIV <- CreateTables(data=train, trt=TREATMENT, y=PURCHASE)
View(NIV$Summary)
NIV$Tables$N_OPEN_REV_ACTS

##------------------------------------------------------------
## NWOE analysis, validation
##------------------------------------------------------------
data(train, package="Information")
data(validation, package="Information")
NIV <- CreateTables(data=train, valid=validation, trt=TREATMENT, y=PURCHASE)
View(NIV$Summary)
NIV$Tables$N_OPEN_REV_ACTS
}

