% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiWOEPlot.R
\name{MultiPlotWOE}
\alias{MultiPlotWOE}
\title{Plot mutiple WOE vectors on one page}
\usage{
MultiPlotWOE(information_object, variables, same_scale = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variables}{vector of variables}
}
\description{
\code{MultiPlotWOE} creates a multiple WOE bar charts on the same page for a specified vector of variables.
Multiple pages will be created if needed (9 plots per page).
}
\examples{
##------------------------------------------------------------
## WOE plots, no validation
##------------------------------------------------------------
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- CreateTables(data=train, y="PURCHASE")
MultiPlotWOE(IV, c("N_OPEN_REV_ACTS", "AGE"))
}

