% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotWOE.R
\name{PlotWOE}
\alias{PlotWOE}
\title{Plot a WOE vector}
\usage{
PlotWOE(information_object, variable, show_values = FALSE)
}
\arguments{
\item{information_object}{object from the information package}

\item{variable}{variable for which we want to see the WOE pattern}

\item{show_values}{do you want to show values on the bar chart? Default is FALSE}
}
\description{
\code{PlotWOE} creates a bar chart of the WOE pattern for a specified variable
}
\examples{
##------------------------------------------------------------
## WOE plot, no validation
##------------------------------------------------------------
data(train, package="Information")
train <- subset(train, TREATMENT==1)
IV <- CreateInfoTables(data=train, y="PURCHASE")
Information::PlotWOE(IV, "N_OPEN_REV_ACTS")
}

